requirements.txt
langchain==0.3.14
langchain-groq==0.2.3
langchain-community==0.3.14
llama-index==0.12.4
llama-index-llms-groq==0.3.1
llama-index-embeddings-huggingface==0.4.0
pinecone==6.0.0
sentence-transformers==3.3.1
streamlit==1.41.1
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.32.3
arxiv==2.1.3
duckduckgo-search==7.2.1
pypdf==5.1.0
pandas==2.2.3
numpy==1.26.4
wikipedia==1.4.0